+++ FRC2018/src/Robot.cpp
@@ -229,7 +229,7 @@ void Robot::TankDrive() {
 	}
 }
 
-void Robot::GyroTurn(bool btn, float speed, double angle) { // Turn based on button value
+void Robot::GyroTurn(bool btn, float speed, float angle) { // Turn based on button value
 	if (btn || gyroTurning) {
 		latestYaw = ahrs->GetYaw();
 		SmartDashboard::PutNumber("Target Angle", angle);
